generic: move candidate state into inference context
